+++ test/browser/features/fixtures/packages/oldest-batch-removed/src/index.js
@@ -4,15 +4,19 @@ const parameters = new URLSearchParams(window.location.search)
 const apiKey = parameters.get('api_key')
 const endpoint = parameters.get('endpoint')
 
-BugsnagPerformance.start({ apiKey, endpoint, retryQueueMaxSize: 3, maximumBatchSize: 1, autoInstrumentFullPageLoads: false, autoInstrumentNetworkRequests: false })
+const sleep = (time) => new Promise((resolve) => setTimeout(resolve, time))
+
+BugsnagPerformance.start({ apiKey, endpoint, retryQueueMaxSize: 3, maximumBatchSize: 1, autoInstrumentFullPageLoads: false, autoInstrumentNetworkRequests: false, appVersion: '1.2.3' })
 
 document.getElementById("send-first-span").addEventListener("click", () => {
   BugsnagPerformance.startSpan("Custom/Span to discard").end()
 })
 
-document.getElementById("send-retry-spans").addEventListener("click", () => {
+document.getElementById("send-retry-spans").addEventListener("click", async () => {
   BugsnagPerformance.startSpan("Custom/Span to retry 1").end()
+  await sleep(100)
   BugsnagPerformance.startSpan("Custom/Span to retry 2").end()
+  await sleep(100)
   BugsnagPerformance.startSpan("Custom/Span to retry 3").end()
 })