+++ test/browser/features/fixtures/packages/traceparent/src/index.js
@@ -24,7 +24,25 @@ document.getElementById("xhr").addEventListener("click", () => {
   xhr.send()
 })
 
-document.getElementById("fetch").addEventListener("click", () => {
+document.getElementById("fetch-simple-headers-in-options").addEventListener("click", () => {
   fetch(reflectEndpoint, { headers: { 'X-Test-Header': 'test' }})
 })
 
+document.getElementById("fetch-headers-class-in-options").addEventListener("click", () => {
+  fetch(reflectEndpoint, { headers: new Headers({ 'X-Test-Header': 'test' })})
+})
+
+document.getElementById("fetch-simple-headers-in-request").addEventListener("click", () => {
+  const request = new Request(reflectEndpoint, { headers: { 'X-Test-Header': 'test' }})
+  fetch(request)
+})
+
+document.getElementById("fetch-headers-class-in-request").addEventListener("click", () => {
+  const request = new Request(reflectEndpoint, { headers: new Headers({ 'X-Test-Header': 'test' })})
+  fetch(request)
+})
+
+document.getElementById("fetch-headers-in-request-and-options").addEventListener("click", () => {
+  const request = new Request(reflectEndpoint, { headers: new Headers({ 'X-Ignored-Header': 'will be ignored' })})
+  fetch(request, { headers: { 'X-Test-Header': 'test' }})
+})